# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CpfConsultation {
  success: boolean;
  cpf: string;
  /** @nullable */
  nome: string | null;
  /** @nullable */
  dataNascimento: string | null;
  source: string;
}

export interface ErrorResponse {
  success: boolean;
  error: string;
}

export type PaymentRequestTrackingParameters = {[key: string]: string};

export interface PaymentRequest {
  name: string;
  email: string;
  cpf: string;
  phone?: string;
  pixKey?: string;
  offerSlug?: string;
  trackingParameters?: PaymentRequestTrackingParameters;
}

export interface PaymentResponse {
  success: boolean;
  status: string;
  /** @nullable */
  transactionId?: string | null;
  /** @nullable */
  pixCode?: string | null;
  /** @nullable */
  qrCodeUrl?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface PaymentStatusResponse {
  success: boolean;
  status: string;
  paid: boolean;
  /** @nullable */
  error?: string | null;
}

export type TrackingConfigResponsePixelsItem = {
  platform: string;
  pixelId: string;
  /** @nullable */
  code: string | null;
  /** @nullable */
  label?: string | null;
};

export interface TrackingConfigResponse {
  success: boolean;
  pixels: TrackingConfigResponsePixelsItem[];
}

export type FunnelEventRequestEventName = typeof FunnelEventRequestEventName[keyof typeof FunnelEventRequestEventName];


export const FunnelEventRequestEventName = {
  funnel_stage_viewed: 'funnel_stage_viewed',
} as const;

export type FunnelEventRequestStage = typeof FunnelEventRequestStage[keyof typeof FunnelEventRequestStage];


export const FunnelEventRequestStage = {
  consulta: 'consulta',
  identidade: 'identidade',
  recebimento: 'recebimento',
} as const;

export interface FunnelEventRequest {
  sessionId: string;
  eventName: FunnelEventRequestEventName;
  stage: FunnelEventRequestStage;
  offerSlug?: string;
}

export type ConsultCpfParams = {
/**
 * @pattern ^[0-9.\-]{11,14}$
 */
cpf: string;
};

export type TrackFunnelEvent200 = {
  success: boolean;
};