# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/event:
    post:
      operationId: trackFunnelEvent
      tags: [analytics]
      summary: Record an anonymous funnel stage view
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FunnelEventRequest"
      responses:
        "200":
          description: Event recorded
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
        "400":
          description: Invalid event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
      required: [success, error]
...
    FunnelEventRequest:
      type: object
      required: [sessionId, eventName, stage]
      properties:
        sessionId:
          type: string
        eventName:
          type: string
          enum: [funnel_stage_viewed]
        stage:
          type: string
          enum: [consulta, identidade, recebimento]
        offerSlug:
          type: string