# lib/api-zod/src/generated/types/funnelEventRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FunnelEventRequestEventName } from './funnelEventRequestEventName';
import type { FunnelEventRequestStage } from './funnelEventRequestStage';

export interface FunnelEventRequest {
  sessionId: string;
  eventName: FunnelEventRequestEventName;
  stage: FunnelEventRequestStage;
  offerSlug?: string;
}